requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.11.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2023.1
distlib==0.3.7
distro==1.9.0
Django==4.2.4
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-cors-headers==4.4.0
django-crispy-forms==2.1
django-filter==24.2
djangorestframework==3.15.2
filelock==3.12.2
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
Markdown==3.6
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.28.1
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
pipenv==2023.8.20
platformdirs==4.1.0
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
requests==2.31.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.6
virtualenv==20.24.3
whitenoise==6.5.0
yarl==1.9.2
